Added home view, slight README changes
