+++ client/src/views/Home/Home.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import logo from '../../assets/logo.svg';
+import './App.css';
+
+function Home() {
+    return (
+        <div className="App">
+            <header className="App-header">
+                <img src={logo} className="App-logo" alt="logo" />
+                <p>
+                    Edit <code>src/App.js</code> and save to reload.
+                </p>
+                <a
+                    className="App-link"
+                    href="https://reactjs.org"
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    Learn React
+                </a>
+            </header>d
+        </div>
+    );
+}
+
+export default Home;
