+++ client/src/views/Home/Home.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Home from './Home';
+
+it('renders without crashing', () => {
+    const div = document.createElement('div');
+    ReactDOM.render(<Home />, div);
+    ReactDOM.unmountComponentAtNode(div);
+});
